requirements.txt
aio-pika==6.8.2
aiofiles==0.8.0
aiohttp==3.8.1
aioredis==2.0.1
aiormq==3.3.1
aiosignal==1.2.0
AMQPStorm==2.10.4
anyio==3.5.0
apyio==0.2.0
asgiref==3.5.0
async-timeout==4.0.2
attrs==21.4.0
autopep8==1.6.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
click==8.0.4
cryptography==36.0.1
ecdsa==0.17.0
fastapi==0.74.0
flake8==4.0.1
frozenlist==1.3.0
h11==0.12.0
httpcore==0.14.7
httpx==0.22.0
hvac==0.11.2
idna==3.3
mccabe==0.6.1
motor==2.3.1
multidict==6.0.2
odmantic==0.3.5
pamqp==2.3.0
pyasn1==0.4.8
pycodestyle==2.8.0
pycparser==2.21
pydantic==1.8.2
pyflakes==2.4.0
PyJWT==2.3.0
pymongo==3.12.3
python-jose==3.3.0
pyws1uem==0.0.10
requests==2.27.1
rfc3986==1.5.0
rsa==4.8
six==1.16.0
sniffio==1.2.0
starlette==0.17.1
stdio-proxy==0.1.3
toml==0.10.2
typing_extensions==4.1.1
urllib3==1.26.8
uvicorn==0.17.5
yarl==1.7.2
